+++ src/test/java/org/eclipse/keyple/core/util/BerTlvUtilTest.java
@@ -27,74 +27,74 @@ public class BerTlvUtilTest {
 
   @Test
   public void parse_whenStructureIsValidAndPrimitiveOnlyIsFalse_shouldProvideAllTags() {
-    Map<Integer, List<byte[]>> tlvs = BerTlvUtil.parse(ByteArrayUtil.fromHex(TLV1), false);
+    Map<Integer, List<byte[]>> tlvs = BerTlvUtil.parse(HexUtil.toByteArray(TLV1), false);
     assertThat(tlvs).containsOnlyKeys(0x6F, 0x84, 0xA5, 0xBF0C, 0x53, 0xC7);
     assertThat(tlvs.get(0x6F))
         .containsExactly(
-            ByteArrayUtil.fromHex(
+            HexUtil.toByteArray(
                 "8409315449432E49434131A516BF0C13C708000000001122334453070A3C2005141001"));
-    assertThat(tlvs.get(0x84)).containsExactly(ByteArrayUtil.fromHex("315449432E49434131"));
+    assertThat(tlvs.get(0x84)).containsExactly(HexUtil.toByteArray("315449432E49434131"));
     assertThat(tlvs.get(0xA5))
-        .containsExactly(ByteArrayUtil.fromHex("BF0C13C708000000001122334453070A3C2005141001"));
+        .containsExactly(HexUtil.toByteArray("BF0C13C708000000001122334453070A3C2005141001"));
     assertThat(tlvs.get(0xBF0C))
-        .containsExactly(ByteArrayUtil.fromHex("C708000000001122334453070A3C2005141001"));
-    assertThat(tlvs.get(0x53)).containsExactly(ByteArrayUtil.fromHex("0A3C2005141001"));
-    assertThat(tlvs.get(0xC7)).containsExactly(ByteArrayUtil.fromHex("0000000011223344"));
+        .containsExactly(HexUtil.toByteArray("C708000000001122334453070A3C2005141001"));
+    assertThat(tlvs.get(0x53)).containsExactly(HexUtil.toByteArray("0A3C2005141001"));
+    assertThat(tlvs.get(0xC7)).containsExactly(HexUtil.toByteArray("0000000011223344"));
   }
 
   @Test
   public void parse_whenStructureIsValidAndPrimitiveOnlyIsFalse_shouldProvideAllTags2() {
     Map<Integer, List<byte[]>> tlvs =
         BerTlvUtil.parse(
-            ByteArrayUtil.fromHex(
+            HexUtil.toByteArray(
                 "E030C106200107021D01C106202009021D04C106206919091D01C106201008041D03C10620401D021D01C10620501E021D01"),
             false);
     assertThat(tlvs).containsOnlyKeys(0xE0, 0xC1);
     assertThat(tlvs.get(0xE0))
         .containsExactly(
-            ByteArrayUtil.fromHex(
+            HexUtil.toByteArray(
                 "C106200107021D01C106202009021D04C106206919091D01C106201008041D03C10620401D021D01C10620501E021D01"));
     assertThat(tlvs.get(0xC1))
         .containsExactly(
-            ByteArrayUtil.fromHex("200107021D01"),
-            ByteArrayUtil.fromHex("202009021D04"),
-            ByteArrayUtil.fromHex("206919091D01"),
-            ByteArrayUtil.fromHex("201008041D03"),
-            ByteArrayUtil.fromHex("20401D021D01"),
-            ByteArrayUtil.fromHex("20501E021D01"));
+            HexUtil.toByteArray("200107021D01"),
+            HexUtil.toByteArray("202009021D04"),
+            HexUtil.toByteArray("206919091D01"),
+            HexUtil.toByteArray("201008041D03"),
+            HexUtil.toByteArray("20401D021D01"),
+            HexUtil.toByteArray("20501E021D01"));
   }
 
   @Test
   public void parseSimple_whenStructureIsValidAndPrimitiveOnlyIsFalse_shouldProvideAllTags() {
-    Map<Integer, byte[]> tlvs = BerTlvUtil.parseSimple(ByteArrayUtil.fromHex(TLV1), false);
+    Map<Integer, byte[]> tlvs = BerTlvUtil.parseSimple(HexUtil.toByteArray(TLV1), false);
     assertThat(tlvs)
         .containsOnly(
             entry(
                 0x6F,
-                ByteArrayUtil.fromHex(
+                HexUtil.toByteArray(
                     "8409315449432E49434131A516BF0C13C708000000001122334453070A3C2005141001")),
-            entry(0x84, ByteArrayUtil.fromHex("315449432E49434131")),
-            entry(0xA5, ByteArrayUtil.fromHex("BF0C13C708000000001122334453070A3C2005141001")),
-            entry(0xBF0C, ByteArrayUtil.fromHex("C708000000001122334453070A3C2005141001")),
-            entry(0x53, ByteArrayUtil.fromHex("0A3C2005141001")),
-            entry(0xC7, ByteArrayUtil.fromHex("0000000011223344")));
+            entry(0x84, HexUtil.toByteArray("315449432E49434131")),
+            entry(0xA5, HexUtil.toByteArray("BF0C13C708000000001122334453070A3C2005141001")),
+            entry(0xBF0C, HexUtil.toByteArray("C708000000001122334453070A3C2005141001")),
+            entry(0x53, HexUtil.toByteArray("0A3C2005141001")),
+            entry(0xC7, HexUtil.toByteArray("0000000011223344")));
   }
 
   @Test
   public void
       parseSimple_whenStructureIsValidAndPrimitiveOnlyIsTrue_shouldProvideOnlyPrimitiveTags() {
-    Map<Integer, byte[]> tlvs = BerTlvUtil.parseSimple(ByteArrayUtil.fromHex(TLV1), true);
+    Map<Integer, byte[]> tlvs = BerTlvUtil.parseSimple(HexUtil.toByteArray(TLV1), true);
     assertThat(tlvs)
         .containsOnly(
-            entry(0x84, ByteArrayUtil.fromHex("315449432E49434131")),
-            entry(0x53, ByteArrayUtil.fromHex("0A3C2005141001")),
-            entry(0xC7, ByteArrayUtil.fromHex("0000000011223344")));
+            entry(0x84, HexUtil.toByteArray("315449432E49434131")),
+            entry(0x53, HexUtil.toByteArray("0A3C2005141001")),
+            entry(0xC7, HexUtil.toByteArray("0000000011223344")));
   }
 
   @Test
   public void parseSimple_whenTagsOrderChange_shouldProvideTheSameTags() {
-    Map<Integer, byte[]> tlvs1 = BerTlvUtil.parseSimple(ByteArrayUtil.fromHex(TLV1), true);
-    Map<Integer, byte[]> tlvs2 = BerTlvUtil.parseSimple(ByteArrayUtil.fromHex(TLV2), true);
+    Map<Integer, byte[]> tlvs1 = BerTlvUtil.parseSimple(HexUtil.toByteArray(TLV1), true);
+    Map<Integer, byte[]> tlvs2 = BerTlvUtil.parseSimple(HexUtil.toByteArray(TLV2), true);
     assertThat(tlvs1).containsExactlyEntriesOf(tlvs2);
   }
 
@@ -102,29 +102,29 @@ public class BerTlvUtilTest {
   public void parseSimple_whenTagsIdIs3Bytes_shouldProvideTheTag() {
     Map<Integer, byte[]> tlvs =
         BerTlvUtil.parseSimple(
-            ByteArrayUtil.fromHex(
+            HexUtil.toByteArray(
                 "6F258409315449432E49434131A518BF0C15DFEF2C08000000001122334453070A3C2005141001"),
             true);
     assertThat(tlvs)
         .containsOnly(
-            entry(0x84, ByteArrayUtil.fromHex("315449432E49434131")),
-            entry(0x53, ByteArrayUtil.fromHex("0A3C2005141001")),
-            entry(0xDFEF2C, ByteArrayUtil.fromHex("0000000011223344")));
+            entry(0x84, HexUtil.toByteArray("315449432E49434131")),
+            entry(0x53, HexUtil.toByteArray("0A3C2005141001")),
+            entry(0xDFEF2C, HexUtil.toByteArray("0000000011223344")));
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void parseSimple_whenStructureIsInvalid_shouldIAE() {
-    BerTlvUtil.parseSimple(ByteArrayUtil.fromHex("6F23A5"), true);
+    BerTlvUtil.parseSimple(HexUtil.toByteArray("6F23A5"), true);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void parseSimple_whenLengthFieldIsInvalid_shouldIAE() {
-    BerTlvUtil.parseSimple(ByteArrayUtil.fromHex("6F83A5"), true);
+    BerTlvUtil.parseSimple(HexUtil.toByteArray("6F83A5"), true);
   }
 
   @Test
   public void parseSimple_whenLengthIsZero_shouldReturnEmptyValue() {
-    Map<Integer, byte[]> tlvs = BerTlvUtil.parseSimple(ByteArrayUtil.fromHex("8400"), false);
+    Map<Integer, byte[]> tlvs = BerTlvUtil.parseSimple(HexUtil.toByteArray("8400"), false);
     assertThat(tlvs.get(0x84)).isEmpty();
   }