+++ CHANGELOG.md
@@ -9,15 +9,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - "CHANGELOG.md" file (issue [eclipse/keyple#6]).
 - CI: Forbid the publication of a version already released (issue [#14]).
 - JSON serializer/deserializer for `short` and `long` data types (issue [#17]).
-- `isHexString` method in `Assert` class.
-- `hexToByte` method in `ByteArrayUtil` class.
-- `hexToShort` method in `ByteArrayUtil` class.
-- `hexToInt` method in `ByteArrayUtil` class.
-- `hexToLong` method in `ByteArrayUtil` class.
-- `toHex` methods for `byte`, `short`, `int` and `long` data types in `ByteArrayUtil` class.
-- `parse` method in `BerTlvUtil` class to parse TLV data structure having several identical TAGs (issue [#19]).
+- `Assert.isHexString` method.
+- `HexUtil.isValid` method.
+- `HexUtil.toByteArray` method.
+- `HexUtil.toByte` method.
+- `HexUtil.toShort` method.
+- `HexUtil.toInt` method.
+- `HexUtil.toLong` method.
+- `HexUtil.toHex` methods for `byte[]`, `byte`, `short`, `int` and `long` data types.
+- `ByteArrayUtil.extractBytes` method.
+- `ByteArrayUtil.extractInt` method.
+- `BerTlvUtil.parse` method to parse TLV data structure having several identical TAGs (issue [#19]).
 ### Changed
 - Optimizations of hexadecimal data conversions.
+### Deprecated
+- `ByteArrayUtil.isValidHexString` method replaced by `HexUtil.isValid` method.
+- `ByteArrayUtil.normalizeHexString` method.
+- `ByteArrayUtil.fromHex` method replaced by `HexUtil.toByteArray` method.
+- `ByteArrayUtil.toHex` method replaced by `HexUtil.toHex` method.
+- `ByteArrayUtil.twoBytesToInt` method replaced by `ByteArrayUtil.extractInt` method.
+- `ByteArrayUtil.twoBytesSignedToInt` method replaced by `ByteArrayUtil.extractInt` method.
+- `ByteArrayUtil.threeBytesToInt` method replaced by `ByteArrayUtil.extractInt` method.
+- `ByteArrayUtil.threeBytesSignedToInt` method replaced by `ByteArrayUtil.extractInt` method.
+- `ByteArrayUtil.fourBytesToInt` method replaced by `ByteArrayUtil.extractInt` method.
 
 ## [2.0.0] - 2021-10-06
 This is the initial release.